small fixes and replacing the older randomcuttree
