+++ Java/core/src/main/java/com/amazon/randomcutforest/tree/RandomCutTree.java
@@ -15,28 +15,30 @@
 
 package com.amazon.randomcutforest.tree;
 
-import static com.amazon.randomcutforest.CommonUtils.*;
-import static com.amazon.randomcutforest.tree.Node.isLeftOf;
+import static com.amazon.randomcutforest.CommonUtils.checkNotNull;
+import static com.amazon.randomcutforest.CommonUtils.checkState;
 
 import java.util.Arrays;
 import java.util.Random;
-import java.util.function.Function;
 
-import com.amazon.randomcutforest.MultiVisitor;
 import com.amazon.randomcutforest.Visitor;
 
 /**
  * A Random Cut Tree is a tree data structure whose leaves represent points
  * inserted into the tree and whose interior nodes represent regions of space
  * defined by Bounding Boxes and Cuts. New nodes and leaves are added to the
- * tree by making random cuts. See {@link #addPoint} for details.
+ * tree by making random cuts.
+ *
+ * This tree is implemented based on pointers and uses he same logic as the
+ * Compact Random Cut trees, providing an alternate implementation to
+ * RandomCutTree corresponding to the initial version.
  *
  * The main use of this class is to be updated with points sampled from a
  * stream, and to define traversal methods. Users can then implement a
  * {@link Visitor} which can be submitted to a traversal method in order to
  * compute a statistic from the tree.
  */
-public class RandomCutTree implements ITree<double[]> {
+public class RandomCutTree extends AbstractRandomCutTree<double[], Node, double[]> {
 
     /**
      * By default, trees will not store sequence indexes.
@@ -48,33 +50,10 @@ public class RandomCutTree implements ITree<double[]> {
      */
     public static final boolean DEFAULT_CENTER_OF_MASS_ENABLED = false;
 
-    /**
-     * The flag to determine if sequence indexes are stored
-     */
-
-    private final boolean storeSequenceIndexesEnabled;
-    /**
-     * The flag to determine if center of mass is enabled
-     */
-
-    private final boolean centerOfMassEnabled;
-    /**
-     * The random number generator used create random cuts.
-     */
-    private final Random random;
-    /**
-     * The root node of the tree.
-     */
-    protected Node root;
-
     protected RandomCutTree(Builder<?> builder) {
-        storeSequenceIndexesEnabled = builder.storeSequenceIndexesEnabled;
-        centerOfMassEnabled = builder.centerOfMassEnabled;
-        if (builder.random != null) {
-            random = builder.random;
-        } else {
-            random = new Random();
-        }
+        super(builder.random, builder.boundingBoxCachingEnabled, builder.centerOfMassEnabled,
+                builder.storeSequenceIndexesEnabled);
+        super.rootIndex = null;
     }
 
     /**
@@ -92,14 +71,14 @@ public class RandomCutTree implements ITree<double[]> {
      * @return a new RandomCutTree with optional arguments set to default values.
      */
     public static RandomCutTree defaultTree(long randomSeed) {
-        return builder().randomSeed(randomSeed).build();
+        return builder().random(new Random(randomSeed)).build();
     }
 
     /**
      * @return a new RandomCutTree with optional arguments set to default values.
      */
     public static RandomCutTree defaultTree() {
-        return builder().build();
+        return builder().random(new Random()).build();
     }
 
     /**
@@ -107,7 +86,7 @@ public class RandomCutTree implements ITree<double[]> {
      *         otherwise.
      */
     public boolean centerOfMassEnabled() {
-        return centerOfMassEnabled;
+        return super.enableCenterOfMass;
     }
 
     /**
@@ -115,61 +94,274 @@ public class RandomCutTree implements ITree<double[]> {
      *         otherwise.
      */
     public boolean storeSequenceIndexesEnabled() {
-        return storeSequenceIndexesEnabled;
+        return super.enableSequenceIndices;
     }
 
-    /**
-     * Return a new {@link Cut}, which is chosen uniformly over the space of
-     * possible cuts for the given bounding box.
-     *
-     * @param random A random number generator
-     * @param box    A bounding box that we want to find a random cut for.
-     * @return A new Cut corresponding to a random cut in the bounding box.
-     */
-    static Cut randomCut(Random random, BoundingBox box) {
-        double rangeSum = box.getRangeSum();
-        checkArgument(rangeSum > 0, "box.getRangeSum() must be greater than 0");
-
-        double breakPoint = random.nextDouble() * rangeSum;
-
-        for (int i = 0; i < box.getDimensions(); i++) {
-            double range = box.getRange(i);
-            if (breakPoint <= range) {
-                double cutValue = box.getMinValue(i) + breakPoint;
-
-                // Random cuts have to take a value in the half-open interval [minValue,
-                // maxValue) to ensure that a
-                // Node has a valid left child and right child.
-                if ((cutValue == box.getMaxValue(i)) && (box.getMinValue(i) < box.getMaxValue(i))) {
-                    cutValue = Math.nextAfter(box.getMaxValue(i), box.getMinValue(i));
-                }
-
-                return new Cut(i, cutValue);
-            }
-            breakPoint -= range;
-        }
-
-        throw new IllegalStateException("The break point did not lie inside the expected range");
+    public RandomCutTree(long seed, boolean enableCache, boolean enableCenterOfMass, boolean enableSequenceIndices) {
+        super(seed, enableCache, enableCenterOfMass, enableSequenceIndices);
+        rootIndex = null;
     }
 
-    /**
-     * Replace a Node in the tree structure. This method replaces oldNode with
-     * newNode as a child of oldNode.getParent().
-     *
-     * @param oldNode The node we are replacing.
-     * @param newNode The new node we are inserting into the tree.
-     */
-    static void replaceNode(Node oldNode, Node newNode) {
+    public RandomCutTree(Random random, boolean enableCache, boolean enableCenterOfMass,
+            boolean enableSequenceIndices) {
+        super(random, enableCache, enableCenterOfMass, enableSequenceIndices);
+        rootIndex = null;
+    }
 
-        Node parent = oldNode.getParent();
-        if (parent != null) {
-            if (parent.getLeftChild() == oldNode) {
-                parent.setLeftChild(newNode);
-            } else {
-                parent.setRightChild(newNode);
+    @Override
+    protected boolean leftOf(double[] point, int cutDimension, double cutValue) {
+        return point[cutDimension] <= cutValue;
+    }
+
+    @Override
+    protected boolean checkEqual(double[] oldPoint, double[] point) {
+        return Arrays.equals(oldPoint, point);
+    }
+
+    @Override
+    protected String toString(double[] doubles) {
+        return Arrays.toString(doubles);
+    }
+
+    @Override
+    void setCachedBox(Node node, AbstractBoundingBox<double[]> savedBox) {
+        node.setBoundingBox((BoundingBox) savedBox);
+    }
+
+    @Override
+    void addToBox(Node node, double[] point) {
+        node.setBoundingBox(node.getBoundingBox().getMergedBox(point));
+    }
+
+    @Override
+    void reComputePointSum(Node node, double[] point) {
+        node.reComputePointSum(point);
+    }
+
+    @Override
+    protected double[] getPoint(Node node) {
+        return node.getLeafPoint();
+    }
+
+    @Override
+    protected INode<Node> getNode(Node node) {
+        return node;
+    }
+
+    @Override
+    AbstractBoundingBox<double[]> getInternalTwoPointBox(double[] first, double[] second) {
+        return new BoundingBox(first, second);
+    }
+
+    @Override
+    BoundingBox getBoundingBoxReflate(Node nodeReference) {
+        if (isLeaf(nodeReference)) {
+            return new BoundingBox(nodeReference.getLeafPoint());
+        }
+        if (nodeReference.getBoundingBox() == null) {
+            return recomputeBox(nodeReference);
+        }
+        return nodeReference.getBoundingBox();
+    }
+
+    @Override
+    BoundingBox recomputeBox(Node nodeReference) {
+        nodeReference.setBoundingBox(getBoundingBoxReflate(nodeReference.getLeftChild())
+                .getMergedBox(getBoundingBoxReflate(nodeReference.getRightChild())));
+        return nodeReference.getBoundingBox();
+    }
+
+    @Override
+    AbstractBoundingBox<double[]> getLeafBoxFromLeafNode(Node node) {
+        return new BoundingBox(node.getLeafPoint());
+    }
+
+    @Override
+    AbstractBoundingBox<double[]> getMutableLeafBoxFromLeafNode(Node node) {
+        double[] leafPoint = node.getLeafPoint();
+        return new BoundingBox(Arrays.copyOf(leafPoint, leafPoint.length), Arrays.copyOf(leafPoint, leafPoint.length),
+                0);
+    }
+
+    // gets the actual point
+    @Override
+    double[] getPointFromLeafNode(Node node) {
+        checkState(node.isLeaf(), "Incorrect use");
+        return node.getLeafPoint();
+    }
+
+    @Override
+    double[] getPointFromPointReference(double[] pointIndex) {
+        return pointIndex;
+    }
+
+    // gets the reference, which happens to be the same point
+    @Override
+    double[] getPointReference(Node node) {
+        return node.getLeafPoint();
+    }
+
+    @Override
+    protected boolean isLeaf(Node node) {
+        return node.isLeaf();
+    }
+
+    @Override
+    protected int decrementMass(Node node) {
+        return node.decrementMass();
+    }
+
+    @Override
+    protected int incrementMass(Node node) {
+        return node.incrementMass();
+    }
+
+    @Override
+    protected Node getSibling(Node node) {
+        Node parent = node.getParent();
+        if (parent.getLeftChild() == node) {
+            return parent.getRightChild();
+        } else if (parent.getRightChild() == node) {
+            return parent.getLeftChild();
+        }
+        throw new IllegalStateException(" incorrect state ");
+    }
+
+    @Override
+    protected Node getParent(Node node) {
+        return node.getParent();
+    }
+
+    @Override
+    protected void setParent(Node node, Node parent) {
+        node.setParent(parent);
+    }
+
+    @Override
+    protected void delete(Node node) {
+    }
+
+    @Override
+    protected int getCutDimension(Node node) {
+        return node.getCutDimension();
+    }
+
+    @Override
+    protected double getCutValue(Node node) {
+        return node.getCut().getValue();
+    }
+
+    @Override
+    protected Node getLeftChild(Node node) {
+        return node.getLeftChild();
+    }
+
+    @Override
+    protected Node getRightChild(Node node) {
+        return node.getRightChild();
+    }
+
+    @Override
+    void replaceNode(Node parent, Node child, Node otherNode) {
+        if (parent.getLeftChild() == child) {
+            parent.setLeftChild(otherNode);
+        } else if (parent.getRightChild() == child) {
+            parent.setRightChild(otherNode);
+        } else {
+            throw new IllegalStateException(" incorrect state ");
+        }
+        otherNode.setParent(parent);
+    }
+
+    @Override
+    protected void replaceNodeBySibling(Node grandParent, Node parent, Node node) {
+        replaceNode(grandParent, parent, getSibling(node));
+    }
+
+    @Override
+    protected Node addLeaf(Node parent, double[] pointIndex, int mass) {
+        Node candidate = new Node(pointIndex);
+        candidate.setMass(mass);
+        candidate.setParent(parent);
+        return candidate;
+    }
+
+    @Override
+    protected Node addNode(Node parent, Node leftChild, Node rightChild, int cutDimension, double cutValue, int mass) {
+        Node candidate = new Node(leftChild, rightChild, new Cut(cutDimension, cutValue), null, enableCenterOfMass);
+        candidate.setParent(parent);
+        candidate.setMass(mass);
+        return candidate;
+    }
+
+    @Override
+    protected void increaseMassOfAncestors(Node mergedNode) {
+        Node parent = mergedNode.getParent();
+        while (parent != null) {
+            parent.incrementMass();
+            parent = parent.getParent();
+        }
+    }
+
+    @Override
+    protected int getMass(Node node) {
+        return node.getMass();
+    }
+
+    protected Node getRoot() {
+        return rootIndex;
+    }
+
+    @Override
+    protected void addSequenceIndex(Node node, long uniqueSequenceNumber) {
+        node.addSequenceIndex(uniqueSequenceNumber);
+    }
+
+    @Override
+    protected void deleteSequenceIndex(Node node, long uniqueSequenceNumber) {
+        if (enableSequenceIndices) {
+            if (!node.getSequenceIndexes().contains(uniqueSequenceNumber)) {
+                throw new IllegalStateException("Error in sequence index. Inconsistency in trees in delete step.");
             }
         }
-        newNode.setParent(parent);
+        node.deleteSequenceIndex(uniqueSequenceNumber);
+    }
+
+    public static class Builder<T extends RandomCutTree.Builder<T>> {
+        private boolean storeSequenceIndexesEnabled = DEFAULT_STORE_SEQUENCE_INDEXES_ENABLED;
+        private boolean centerOfMassEnabled = DEFAULT_CENTER_OF_MASS_ENABLED;
+        private boolean boundingBoxCachingEnabled = true;
+        private Random random;
+
+        public T storeSequenceIndexesEnabled(boolean storeSequenceIndexesEnabled) {
+            this.storeSequenceIndexesEnabled = storeSequenceIndexesEnabled;
+            return (T) this;
+        }
+
+        public T centerOfMassEnabled(boolean centerOfMassEnabled) {
+            this.centerOfMassEnabled = centerOfMassEnabled;
+            return (T) this;
+        }
+
+        public T enableBoundingBoxCaching(boolean boundingBoxCacheEnabled) {
+            this.boundingBoxCachingEnabled = boundingBoxCacheEnabled;
+            return (T) this;
+        }
+
+        public T randomSeed(long randomSeed) {
+            this.random = new Random(randomSeed);
+            return (T) this;
+        }
+
+        public T random(Random random) {
+            this.random = random;
+            return (T) this;
+        }
+
+        public RandomCutTree build() {
+            return new RandomCutTree(random, boundingBoxCachingEnabled, centerOfMassEnabled,
+                    storeSequenceIndexesEnabled);
+        }
     }
 
     /**
@@ -179,7 +371,7 @@ public class RandomCutTree implements ITree<double[]> {
      * @param node The node whose sibling we are requesting.
      * @return the sibling of node in the tree.
      */
-    static Node getSibling(Node node) {
+    static Node getSiblingStatic(Node node) {
         checkNotNull(node.getParent(), "node parent must not be null");
 
         Node parent = node.getParent();
@@ -192,351 +384,4 @@ public class RandomCutTree implements ITree<double[]> {
         }
     }
 
-    /**
-     * Delete the given point from this tree.
-     *
-     * @param point         A point in the tree that we wish to delete.
-     * @param sequenceIndex The sequence index of the point.
-     */
-    @Override
-    public void deletePoint(double[] point, long sequenceIndex) {
-        checkState(root != null, "root must not be null");
-        deletePoint(root, point, sequenceIndex);
-    }
-
-    /**
-     * This function deletes the point from the tree recursively. We traverse the
-     * tree based on the cut stored in each interior node until we reach a leaf
-     * node. We then delete the leaf node if the mass of the node is 1, otherwise we
-     * reduce the mass by 1.
-     *
-     * @param node          node that we are visiting in the tree.
-     * @param point         the point that is being deleted from the tree.
-     * @param sequenceIndex the insertion index of the point being deleted from the
-     *                      tree.
-     */
-    void deletePoint(Node node, double[] point, long sequenceIndex) {
-        if (node.isLeaf()) {
-            if (!node.leafPointEquals(point)) {
-                throw new IllegalStateException(Arrays.toString(point) + " " + Arrays.toString(node.getLeafPoint())
-                        + " " + Arrays.equals(node.getLeafPoint(), point) + " Inconsistency in trees in delete step.");
-            }
-            if (storeSequenceIndexesEnabled) {
-                if (!node.getSequenceIndexes().contains(sequenceIndex)) {
-                    throw new IllegalStateException("Error in sequence index. Inconsistency in trees in delete step.");
-                }
-            }
-
-            /*
-             * the above assumes that sequence indexes are unique ... which is true for the
-             * specific sampler used
-             */
-
-            if (node.getMass() > 1) {
-                node.decrementMass();
-                if (storeSequenceIndexesEnabled) {
-                    node.deleteSequenceIndex(sequenceIndex);
-                }
-                return;
-            }
-
-            Node parent = node.getParent();
-            if (parent == null) {
-                root = null;
-                return;
-            }
-
-            Node grandParent = parent.getParent();
-            if (grandParent == null) {
-                root = getSibling(node);
-                root.setParent(null);
-            } else {
-                Node sibling = getSibling(node);
-                replaceNode(parent, sibling);
-            }
-            return;
-        }
-
-        // node is not a leaf
-
-        if (isLeftOf(point, node)) {
-            deletePoint(node.getLeftChild(), point, sequenceIndex);
-        } else {
-            deletePoint(node.getRightChild(), point, sequenceIndex);
-        }
-
-        BoundingBox leftBox = node.getLeftChild().getBoundingBox();
-        BoundingBox rightBox = node.getRightChild().getBoundingBox();
-        node.setBoundingBox(leftBox.getMergedBox(rightBox));
-        node.decrementMass();
-        if (centerOfMassEnabled) {
-            node.subtractFromPointSum(point);
-        }
-    }
-
-    /**
-     * Add a new point to the tree.
-     *
-     * @param point The point to add to the tree.
-     */
-    @Override
-    public double[] addPoint(double[] point, long sequenceIndex) {
-        if (root == null) {
-            root = newLeafNode(point, sequenceIndex);
-        } else {
-            addPoint(root, point, sequenceIndex);
-        }
-        return point;
-    }
-
-    /**
-     * This function adds a point to the tree recursively. The algorithm for adding
-     * a point is as follows:
-     * <ol>
-     * <li>At the current node we create a new bounding box by merging the point
-     * with the existing box.</li>
-     * <li>We pick a dimension and a value of cut.</li>
-     * <li>If the cut falls outside the existing box we create a new node and
-     * replace the current node with the new node. We then add the current node as
-     * the child node to the new node. We create another leaf node containing the
-     * point that we want to insert and add it as the other child of the node that
-     * we have created earlier.</li>
-     * <li>If the cut falls inside the existing box. We follow the cut of the
-     * existing box and move to right or left child of the current node based on the
-     * existing cut.</li>
-     * </ol>
-     *
-     * @param node          the current node in the tree we are on
-     * @param point         the point that we want to add to the tree
-     * @param sequenceIndex the insertion index of the point being added to the tree
-     */
-
-    private void addPoint(Node node, double[] point, long sequenceIndex) {
-        if (node.isLeaf() && node.leafPointEquals(point)) {
-            // the inserted point is equal to an existing leaf point
-            node.incrementMass();
-            if (storeSequenceIndexesEnabled) {
-                node.addSequenceIndex(sequenceIndex);
-            }
-            return;
-        }
-
-        // either the node is not a leaf, or else it's a leaf node containing a
-        // different point
-
-        BoundingBox existingBox = node.getBoundingBox();
-        BoundingBox mergedBox = existingBox.getMergedBox(point);
-
-        if (!existingBox.contains(point)) {
-
-            // Propose a random cut
-
-            Cut cut = randomCut(random, mergedBox);
-            int splitDimension = cut.getDimension();
-            double splitValue = cut.getValue();
-            double minValue = existingBox.getMinValue(splitDimension);
-            double maxValue = existingBox.getMaxValue(splitDimension);
-
-            // if the proposed cut separates the new point from the existing bounding box:
-            // * create a new leaf node for the point
-            // * make it a sibling of the existing bounding box
-            // * make the new leaf node and the existing node children of a new node with
-            // the merged bounding box
-
-            if (minValue > splitValue || maxValue <= splitValue) {
-                Node leaf = newLeafNode(point, sequenceIndex);
-                Node mergedNode = minValue > splitValue ? newNode(leaf, node, cut, mergedBox)
-                        : newNode(node, leaf, cut, mergedBox);
-                if (node.getParent() == null) {
-                    root = mergedNode;
-                } else {
-                    replaceNode(node, mergedNode);
-                }
-                leaf.setParent(mergedNode);
-                node.setParent(mergedNode);
-
-                return;
-            }
-        }
-
-        // Either the new point is contained in this node's bounding box, or else the
-        // proposed cut did not separate
-        // it from the existing bounding box. Try again at the next level.
-
-        if (isLeftOf(point, node)) {
-            addPoint(node.getLeftChild(), point, sequenceIndex);
-        } else {
-            addPoint(node.getRightChild(), point, sequenceIndex);
-        }
-
-        node.setBoundingBox(mergedBox);
-        node.incrementMass();
-        if (centerOfMassEnabled) {
-            node.addToPointSum(point);
-        }
-    }
-
-    /*
-     * In the method below, the newCopy() operator is invoked after traversing the
-     * left side by design. It is not currently being used; but in theory this
-     * allows inorder traversal which can simulate more functions in comparison to
-     * post-order traversal. One can (but would not be advised to) use the
-     * properties of the left traversal to inform the right traversal. Such a
-     * strategy may be useful in correcting known biases (for example, the algorithm
-     * is biased lower etc.) but in most cases the newCopy() should just use the
-     * size information of the data structure and copy the information that has been
-     * passed to it from its parent.
-     */
-
-    /**
-     * Starting from the root, traverse the canonical path to a leaf node and visit
-     * the nodes along the path. The canonical path is determined by the input
-     * point: at each interior node, we select the child node by comparing the
-     * node's {@link Cut} to the corresponding coordinate value in the input point.
-     * The method recursively traverses to the leaf node first and then invokes the
-     * visitor on each node in reverse order. That is, if the path to the leaf node
-     * determined by the input point is root, node1, node2, ..., node(N-1), nodeN,
-     * leaf; then we will first invoke visitor::acceptLeaf on the leaf node, and
-     * then we will invoke visitor::accept on the remaining nodes in the following
-     * order: nodeN, node(N-1), ..., node2, node1, and root.
-     *
-     * @param point          A point which determines the traversal path from the
-     *                       root to a leaf node.
-     * @param visitorFactory A visitor that will be invoked for each node on the
-     *                       path.
-     * @param <R>            The return type of the Visitor.
-     * @return the value of {@link Visitor#getResult()}} after the traversal.
-     */
-    @Override
-    public <R> R traverse(double[] point, Function<ITree<?>, Visitor<R>> visitorFactory) {
-        checkState(root != null, "this tree doesn't contain any nodes");
-        Visitor<R> visitor = visitorFactory.apply(this);
-        traversePathToLeafAndVisitNodes(point, visitor, root, 0);
-        return visitor.getResult();
-    }
-
-    private <R> void traversePathToLeafAndVisitNodes(double[] point, Visitor<R> visitor, Node currentNode,
-            int depthOfNode) {
-        if (currentNode.isLeaf()) {
-            visitor.acceptLeaf(currentNode, depthOfNode);
-        } else {
-            Node childNode = isLeftOf(point, currentNode) ? currentNode.getLeftChild() : currentNode.getRightChild();
-            traversePathToLeafAndVisitNodes(point, visitor, childNode, depthOfNode + 1);
-            visitor.accept(currentNode, depthOfNode);
-        }
-    }
-
-    /**
-     * This is a traversal method which follows the standard traveral path (defined
-     * in {@link #traverse(double[], Function)}) but at Node in checks to see
-     * whether the visitor should split. If a split is triggered, then independent
-     * copies of the visitor are sent down each branch of the tree and then merged
-     * before propogating the result.
-     *
-     * @param point          A point which determines the traversal path from the
-     *                       root to a leaf node.
-     * @param visitorFactory A visitor that will be invoked for each node on the
-     *                       path.
-     * @param <R>            The return type of the Visitor.
-     * @return the value of {@link Visitor#getResult()}} after the traversal.
-     */
-    @Override
-    public <R> R traverseMulti(double[] point, Function<ITree<?>, MultiVisitor<R>> visitorFactory) {
-        checkNotNull(point, "point must not be null");
-        checkNotNull(visitorFactory, "visitor must not be null");
-        checkState(root != null, "this tree doesn't contain any nodes");
-        MultiVisitor<R> visitor = visitorFactory.apply(this);
-        traverseTreeMulti(point, visitor, root, 0);
-        return visitor.getResult();
-    }
-
-    private <R> void traverseTreeMulti(double[] point, MultiVisitor<R> visitor, Node currentNode, int depthOfNode) {
-        if (currentNode.isLeaf()) {
-            visitor.acceptLeaf(currentNode, depthOfNode);
-        } else if (visitor.trigger(currentNode)) {
-            traverseTreeMulti(point, visitor, currentNode.getLeftChild(), depthOfNode + 1);
-            MultiVisitor<R> newVisitor = visitor.newCopy();
-            traverseTreeMulti(point, newVisitor, currentNode.getRightChild(), depthOfNode + 1);
-            visitor.combine(newVisitor);
-            visitor.accept(currentNode, depthOfNode);
-        } else {
-            Node childNode = isLeftOf(point, currentNode) ? currentNode.getLeftChild() : currentNode.getRightChild();
-            traverseTreeMulti(point, visitor, childNode, depthOfNode + 1);
-            visitor.accept(currentNode, depthOfNode);
-        }
-    }
-
-    private Node newLeafNode(double[] point, long sequenceIndex) {
-        Node node = new Node(point);
-        node.setMass(1);
-        if (storeSequenceIndexesEnabled) {
-            node.addSequenceIndex(sequenceIndex);
-        }
-        return node;
-    }
-
-    private Node newNode(Node leftChild, Node rightChild, Cut cut, BoundingBox box) {
-        Node node = new Node(leftChild, rightChild, cut, box, centerOfMassEnabled);
-
-        if (leftChild != null) {
-            node.addMass(leftChild.getMass());
-            if (centerOfMassEnabled) {
-                node.addToPointSum(leftChild.getPointSum());
-            }
-        }
-
-        if (rightChild != null) {
-            node.addMass(rightChild.getMass());
-            if (centerOfMassEnabled) {
-                node.addToPointSum(rightChild.getPointSum());
-            }
-        }
-
-        return node;
-    }
-
-    /**
-     * @return the root node in the tree
-     */
-    public Node getRoot() {
-        return root;
-    }
-
-    /**
-     * @return the total mass in the tree.
-     */
-    @Override
-    public int getMass() {
-        return root == null ? 0 : root.getMass();
-    }
-
-    public static class Builder<T extends Builder<T>> {
-        private boolean storeSequenceIndexesEnabled = DEFAULT_STORE_SEQUENCE_INDEXES_ENABLED;
-        private boolean centerOfMassEnabled = DEFAULT_CENTER_OF_MASS_ENABLED;
-        private Random random = null;
-
-        public T storeSequenceIndexesEnabled(boolean storeSequenceIndexesEnabled) {
-            this.storeSequenceIndexesEnabled = storeSequenceIndexesEnabled;
-            return (T) this;
-        }
-
-        public T centerOfMassEnabled(boolean centerOfMassEnabled) {
-            this.centerOfMassEnabled = centerOfMassEnabled;
-            return (T) this;
-        }
-
-        public T random(Random random) {
-            this.random = random;
-            return (T) this;
-        }
-
-        public T randomSeed(long randomSeed) {
-            random = new Random(randomSeed);
-            return (T) this;
-        }
-
-        public RandomCutTree build() {
-            return new RandomCutTree(this);
-        }
-    }
 }
